pyproject.toml
[tool.poetry]
name = "portalprocessos"
version = "0.1.0"
description = "A web application to hold BPMN documentation"
authors = ["André Carvalho <alac1984@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
sqlmodel = "^0.0.8"
alembic = "^1.11.1"
jinja2 = "^3.1.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
aiofiles = "^23.1.0"

[tool.poetry.dev-dependencies]
pynvim = "^0.4.3"
pytest = "^7.3.1"
coverage = "^7.2.7"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
mypy = "^1.3.0"
black = "^23.3.0"
pre-commit = "^3.3.2"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^4.1.0"
pytest-mock = "^3.10.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
aioconsole = "^0.6.1"

[tool.mypy]
exclude = [
    'migrations/'
]

[tool.mypy."models.*"]

[tool.flake8]
max-line-length = 88
exclude = [
    'migrations/*'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
